requirements.txt
# Core Vision & ML
opencv-python>=4.8.0
mediapipe>=0.10.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0

# Streamlit App
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
aiortc>=1.6.0

# AI / LLM
openai>=1.12.0

# Text-to-Speech
gTTS>=2.4.0
pyttsx3>=2.90

# Utilities
python-dotenv>=1.0.0
pillow>=10.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
